"Add forget password functionality using HttpClient and Observable"
This diff adds the ability to send an OTP to a user's email address for password recovery. It introduces the use of HttpClient and Observable to make HTTP requests to a server. The `sendOtp` method is added to the component, which sends a POST request to the server with the user's email address. If the request is successful, the OTP is sent and the user is navigated to the OTP verification page. If the request fails, an error message is displayed. The `HttpClientModule` is also imported to enable the use of HttpClient.