"Added HTTP client for API calls in ConformPasswordComponent and VerifyOtpComponent"
This diff adds the HttpClient module and makes necessary changes to use it for API calls in ConformPasswordComponent and VerifyOtpComponent. The conformPassword and verifyOtp methods are updated to make HTTP POST requests and handle errors using RxJS operators. The SettlerService is also injected in VerifyOtpComponent to access the email object.