"Add addUser, deleteUser, and addShop options to dashboard side menu and corresponding components to router"
This diff adds three new options to the side menu in the dashboard component: "Add User", "Delete User", and "Add Shop". It also adds three new components, AddCustomerComponent, DeleteCustomerComponent, and AddShopComponent, to the router configuration for these new options. The dashboard.component.scss file has a minor change to set the height of the dashboard container to "fit-content" instead of a fixed height. The dashboard.component.ts file has been updated to include methods for navigating to the new components when the corresponding side menu options are clicked.