"Add localStorage id property in AddShopComponent and set id in LoginComponent"
This commit includes changes to two files: add-shop.component.ts and login.component.ts.

In add-shop.component.ts, a new property 'id' is added to the 'body' object, which is set to the value of localStorage.getItem('id'). There is also a console.log statement with a semicolon added for proper formatting.

In login.component.ts, the 'data' object is destructured to extract the 'id' property, which is then set in localStorage using localStorage.setItem('id', id). This occurs within the subscribe block for the login API call.